refactors modal body into a separate component
